ST-384: Remove commented out code and move model to own folder/file
